Handle all discovery options in rmw_context.cpp Signed-off-by: Andrea Sorbini <asorbini@rti.com>
