+++ rmw_connextdds_common/src/common/rmw_context.cpp
@@ -25,6 +25,8 @@
 
 #include "rcutils/env.h"
 #include "rcutils/filesystem.h"
+#include "rcutils/process.h"
+#include "rcutils/snprintf.h"
 
 /******************************************************************************
  * Global reference to the Domain Participant Factory.
@@ -75,6 +77,185 @@ rmw_connextdds_initialize_participant_factory_qos(
 }
 
 static
+rmw_ret_t rmw_connextdds_extend_initial_peer_list(
+  const rmw_peer_address_t * const static_peers,
+  const size_t static_peer_count,
+  struct DDS_StringSeq * const out)
+{
+  if (static_peer_count == 0) {
+    return RMW_RET_OK;
+  }
+
+  if (static_peers == nullptr) {
+    RMW_CONNEXT_LOG_ERROR_A_SET(
+      "received nullptr static_peers but a static_peer_count of %lu",
+      static_peer_count);
+    return RMW_RET_ERROR;
+  }
+
+  const auto initial_length = DDS_StringSeq_get_length(out);
+  const auto new_seq_length = initial_length + static_peer_count;
+  if (!DDS_StringSeq_ensure_length(out, new_seq_length, new_seq_length)) {
+    RMW_CONNEXT_LOG_ERROR_SET("failed to resize string sequence")
+    return RMW_RET_ERROR;
+  }
+  for (size_t s = 0; s < static_peer_count; ++s) {
+    const auto index = initial_length + s;
+    const char * peer = static_peers[s].peer_address;
+    char ** const element_ref = DDS_StringSeq_get_reference(out, index);
+    RMW_CONNEXT_ASSERT(nullptr != element_ref);
+    if (nullptr != *element_ref) {
+      /* If some strings are still hanging around this dead space in the seq,
+         then free their memory. */
+      DDS_String_free(*element_ref);
+    }
+    *element_ref = DDS_String_dup(peer);
+    if (nullptr == *element_ref) {
+      RMW_CONNEXT_LOG_ERROR_A_SET("failed to duplicate peer string: %s", peer);
+      return RMW_RET_ERROR;
+    }
+
+    RMW_CONNEXT_LOG_TRACE_A(
+      "inserted static peer: i=%d, peer='%s'",
+      index, peer);
+  }
+
+  return RMW_RET_OK;
+}
+
+static
+rmw_ret_t
+rmw_connextdds_initialize_discovery_options(
+  rmw_context_impl_t * const ctx,
+  DDS_DomainParticipantQos & dp_qos)
+{
+  const auto range = ctx->discovery_options->automatic_discovery_range;
+  switch (range) {
+    case RMW_AUTOMATIC_DISCOVERY_RANGE_SYSTEM_DEFAULT:
+    case RMW_AUTOMATIC_DISCOVERY_RANGE_SUBNET:
+      /* No action needed. This is the default discovery behavior for DDS */
+      break;
+    case RMW_AUTOMATIC_DISCOVERY_RANGE_NOT_SET:
+    case RMW_AUTOMATIC_DISCOVERY_RANGE_LOCALHOST:
+    case RMW_AUTOMATIC_DISCOVERY_RANGE_OFF:
+      /*Limit the UDPv4 transport to use only unicast interfaces, by setting
+        the list of multicast interfaces to a "localhost" value.
+        Note: We allow the LOCALHOST interface for the OFF range
+        because if we leave this property completely blank then it
+        has the opposite effect and allows all interfaces to be used.
+        Allowing only LOCALHOST at least minimizes the unnecessary
+        discovery traffic and prevents discovery with other host
+        machines, while the domain_tag protects against same-host
+        connections. */
+      if (DDS_RETCODE_OK != DDS_PropertyQosPolicyHelper_assert_property(
+          &dp_qos.property,
+          "dds.transport.UDPv4.builtin.parent.allow_multicast_interfaces_list",
+          RMW_CONNEXT_LOCALHOST_ONLY_ADDRESS,
+          DDS_BOOLEAN_FALSE /* propagate */))
+      {
+        RMW_CONNEXT_LOG_ERROR_A_SET(
+          "failed to assert property on participant: %s",
+          "dds.transport.UDPv4.builtin.parent.allow_multicast_interfaces_list")
+        return RMW_RET_ERROR;
+      }
+      break;
+    default:
+      RMW_CONNEXT_LOG_ERROR_A_SET(
+        "Unknown value provided for automatic discovery range: %i",
+        range);
+      return RMW_RET_ERROR;
+  }
+
+  if (RMW_AUTOMATIC_DISCOVERY_RANGE_OFF == range) {
+    // When discovery rage is RMW_AUTOMATIC_DISCOVERY_RANGE_OFF,
+    // prevent the participant from discovery anyone by setting an empty
+    // initial peers list and disabling "accept_unknown_peers".
+    // Also, assign a host-wide unique domain tag to the participant to
+    // prevent discovery with other local participant (e.g. through shared
+    // memory transport).
+    const DDS_Long ros_peers = DDS_StringSeq_get_length(&ctx->initial_peers);
+    const DDS_Long qos_peers = DDS_StringSeq_get_length(&dp_qos.discovery.initial_peers);
+    dp_qos.discovery.accept_unknown_peers = DDS_BOOLEAN_FALSE;
+    if (ros_peers > 0) {
+      RMW_CONNEXT_LOG_WARNING_A(
+        "requested %d initial peers using %s, but discovery range is off",
+        ros_peers,
+        RMW_CONNEXT_ENV_INITIAL_PEERS);
+      if (!DDS_StringSeq_ensure_length(&ctx->initial_peers, 0, 0)) {
+        RMW_CONNEXT_LOG_ERROR_SET("failed to clear initial peers list")
+        return RMW_RET_ERROR;
+      }
+    }
+    if (qos_peers > 0) {
+      RMW_CONNEXT_LOG_WARNING_A(
+        "requested %d initial peers from DomainParticipantQos, but discovery range is off",
+        qos_peers);
+      if (!DDS_StringSeq_ensure_length(&ctx->initial_peers, 0, 0)) {
+        RMW_CONNEXT_LOG_ERROR_SET("failed to clear initial peers list")
+        return RMW_RET_ERROR;
+      }
+    }
+    /* See earlier note about why we allow LOCALHOST interface for
+        OFF range. */
+    if (DDS_RETCODE_OK != DDS_PropertyQosPolicyHelper_assert_property(
+        &dp_qos.property,
+        "dds.transport.UDPv4.builtin.parent.allow_interfaces_list",
+        RMW_CONNEXT_LOCALHOST_ONLY_ADDRESS,
+        DDS_BOOLEAN_FALSE /* propagate */))
+    {
+      RMW_CONNEXT_LOG_ERROR_SET(
+        "failed to assert property on participant: "
+        "dds.transport.UDPv4.builtin.parent.allow_interfaces_list");
+      return RMW_RET_ERROR;
+    }
+
+    /* Give this participant its own unique domain tag to prevent
+        unicast discovery from happening. */
+    if (!ctx->domain_tag) {
+      const auto pid = rcutils_get_pid();
+      static const char * format_string = "ros_discovery_off_%d";
+      const int bytes_needed = rcutils_snprintf(nullptr, 0, format_string, pid);
+      ctx->domain_tag = DDS_String_alloc(bytes_needed);
+      if (nullptr == ctx->domain_tag) {
+        RMW_CONNEXT_LOG_ERROR_SET("failed to allocate domain tag string");
+        return RMW_RET_BAD_ALLOC;
+      }
+      if (rcutils_snprintf(ctx->domain_tag, bytes_needed + 1, format_string, pid) < 0) {
+        RMW_CONNEXT_LOG_ERROR_SET("failed to format ros discovery off information into domain tag");
+        return RMW_RET_ERROR;
+      }
+    }
+    if (DDS_RETCODE_OK != DDS_PropertyQosPolicyHelper_assert_property(
+        &dp_qos.property,
+        "dds.domain_participant.domain_tag",
+        ctx->domain_tag,
+        DDS_BOOLEAN_FALSE))
+    {
+      RMW_CONNEXT_LOG_ERROR_SET(
+        "failed to assert property on participant: "
+        "dds.domain_participant.domain_tag");
+      return RMW_RET_ERROR;
+    }
+  } else {
+    // For any other discovery range, copy the list of static peers to so that
+    // it will be later copied to DomainParticipantQos::discovery::initial_peers.
+    dp_qos.discovery.accept_unknown_peers = DDS_BOOLEAN_TRUE;
+    const auto rc = rmw_connextdds_extend_initial_peer_list(
+      ctx->discovery_options->static_peers,
+      ctx->discovery_options->static_peers_count,
+      &ctx->initial_peers);
+    if (RMW_RET_OK != rc) {
+      RMW_CONNEXT_LOG_ERROR(
+        "failed to extend initial peers with the static peers");
+      return rc;
+    }
+  }
+
+  return RMW_RET_OK;
+}
+
+
+static
 rmw_ret_t
 rmw_connextdds_initialize_participant_qos(
   rmw_context_impl_t * const ctx,
@@ -97,20 +278,22 @@ rmw_connextdds_initialize_participant_qos(
   switch (ctx->participant_qos_override_policy) {
     case rmw_context_impl_t::participant_qos_override_policy_t::All:
     case rmw_context_impl_t::participant_qos_override_policy_t::Basic:
-      {
-        if (DDS_StringSeq_get_length(&ctx->initial_peers) > 0)
-        {
-          if (!DDS_StringSeq_copy(&dp_qos.discovery.initial_peers, &ctx->initial_peers)) {
-            RMW_CONNEXT_LOG_ERROR_SET("failed to copy initial peers sequence")
-            return RMW_RET_ERROR;
-          }
+      if (nullptr != ctx->discovery_options) {
+        const auto rc = rmw_connextdds_initialize_discovery_options(ctx, dp_qos);
+        if (RMW_RET_OK != rc) {
+          RMW_CONNEXT_LOG_ERROR("failed to initialize discovery options")
+          return RMW_RET_ERROR;
         }
-        break;
       }
-    default:
+      if (DDS_StringSeq_get_length(&ctx->initial_peers) > 0 &&
+        !DDS_StringSeq_copy(&dp_qos.discovery.initial_peers, &ctx->initial_peers))
       {
-        break;
+        RMW_CONNEXT_LOG_ERROR_SET("failed to copy initial peers sequence")
+        return RMW_RET_ERROR;
       }
+      break;
+    default:
+      break;
   }
 
   return RMW_RET_OK;
@@ -122,7 +305,9 @@ rmw_context_impl_t::initialize_node(
 {
   if (this->node_count > 0) {
     bool params_equal = false;
-    if (rmw_discovery_options_equal(this->discovery_options, discovery_options, &params_equal) != RMW_RET_OK) {
+    if (rmw_discovery_options_equal(
+        this->discovery_options, discovery_options, &params_equal) != RMW_RET_OK)
+    {
       RMW_CONNEXT_LOG_ERROR_SET("invalid discovery params argument");
       return RMW_RET_INVALID_ARGUMENT;
     }
@@ -135,10 +320,10 @@ rmw_context_impl_t::initialize_node(
   } else {
     if (nullptr != discovery_options) {
       RMW_CONNEXT_ASSERT(nullptr == this->discovery_options)
-      this->discovery_options = (rmw_discovery_options_t*)
+      this->discovery_options = static_cast<rmw_discovery_options_t *>(
         this->base->options.allocator.allocate(
-        sizeof(rmw_discovery_options_t),
-        this->base->options.allocator.state);
+          sizeof(rmw_discovery_options_t),
+          this->base->options.allocator.state));
       if (nullptr == this->discovery_options) {
         RMW_CONNEXT_LOG_ERROR_SET("failed to allocate discovery options")
         return RMW_RET_ERROR;
@@ -167,30 +352,30 @@ rmw_context_impl_t::initialize_node(
       if (RMW_AUTOMATIC_DISCOVERY_RANGE_LOCALHOST ==
         this->discovery_options->automatic_discovery_range)
       {
-        static const char * const NDDS_DISCOVERY_PEERS = "NDDS_DISCOVERY_PEERS";
+        static const char * const ndds_disc_peers_var = "NDDS_DISCOVERY_PEERS";
         const char * ndds_disc_peers = nullptr;
-        const char * lookup_rc = rcutils_get_env(NDDS_DISCOVERY_PEERS, &ndds_disc_peers);
+        const char * lookup_rc = rcutils_get_env(ndds_disc_peers_var, &ndds_disc_peers);
         if (nullptr != lookup_rc || nullptr == ndds_disc_peers) {
           RMW_CONNEXT_LOG_ERROR_A_SET(
             "failed to lookup from environment: "
             "var=%s, "
             "rc=%s ",
-            NDDS_DISCOVERY_PEERS,
+            ndds_disc_peers_var,
             lookup_rc)
           return RMW_RET_ERROR;
         }
-        if (ndds_disc_peers[0] == '\0'
-          && !rcutils_set_env(NDDS_DISCOVERY_PEERS, RMW_CONNEXT_LOCALHOST_ONLY_ADDRESS))
+        if (ndds_disc_peers[0] == '\0' &&
+          !rcutils_set_env(ndds_disc_peers_var, RMW_CONNEXT_LOCALHOST_ONLY_ADDRESS))
         {
           RMW_CONNEXT_LOG_ERROR_A_SET(
-            "failed to set environment: var=%s",
-            NDDS_DISCOVERY_PEERS)
+            "failed to set environment: var=%s", ndds_disc_peers_var)
           return RMW_RET_ERROR;
         }
       }
     }
 
     if (nullptr == RMW_Connext_gv_DomainParticipantFactory) {
+      RMW_CONNEXT_ASSERT(1 == RMW_Connext_gv_ContextCount)
       RMW_CONNEXT_LOG_DEBUG("initializing DDS DomainParticipantFactory")
 
       if (RMW_RET_OK !=
