+++ rmw_connextdds_common/src/common/rmw_impl.cpp
@@ -205,54 +205,6 @@ rmw_connextdds_parse_string_list(
   return RMW_RET_OK;
 }
 
-rmw_ret_t rmw_connextdds_extend_initial_peer_list(
-  const rmw_peer_address_t * const static_peers,
-  const size_t static_peer_count,
-  struct DDS_StringSeq * const out)
-{
-  if (static_peer_count == 0) {
-    return RMW_RET_OK;
-  }
-
-  if (static_peers == nullptr) {
-    RMW_CONNEXT_LOG_ERROR_A_SET(
-      "received nullptr static_peers but a static_peer_count of %lu",
-      static_peer_count);
-    return RMW_RET_ERROR;
-  }
-
-  const auto initial_length = DDS_StringSeq_get_length(out);
-  const auto new_seq_length = initial_length + static_peer_count;
-  if (!DDS_StringSeq_ensure_length(out, new_seq_length, new_seq_length)) {
-    RMW_CONNEXT_LOG_ERROR_SET("failed to resize string sequence")
-    return RMW_RET_ERROR;
-  }
-  for (size_t s=0; s < static_peer_count; ++s) {
-    const auto index = initial_length + s;
-    const char * peer = static_peers[s].peer_address;
-    char ** const element_ref = DDS_StringSeq_get_reference(out, index);
-    RMW_CONNEXT_ASSERT(nullptr != element_ref);
-    if (nullptr != *element_ref) {
-      /* If some strings are still hanging around this dead space in the seq,
-         then free their memory. */
-      DDS_String_free(*element_ref);
-    }
-    const auto peer_char_len = strnlen(peer, RMW_DISCOVERY_OPTIONS_STATIC_PEERS_MAX_LENGTH);
-    *element_ref = DDS_String_alloc(peer_char_len + 1);  // +1 for nul character
-    if (nullptr == *element_ref) {
-      RMW_CONNEXT_LOG_ERROR_SET("failed to allocate string");
-      return RMW_RET_ERROR;
-    }
-
-    std::memcpy(*element_ref, peer, peer_char_len + 1);
-    RMW_CONNEXT_LOG_TRACE_A(
-      "inserted static peer: i=%d, peer='%s'",
-      index, peer);
-  }
-
-  return RMW_RET_OK;
-}
-
 bool
 rmw_connextdds_find_string_in_list(
   const DDS_StringSeq * const values,
